+++ test/fixtures/base/main.tf
@@ -15,7 +15,6 @@
 module "workspace" {
   source = "../../../"
 
-
   description       = "Test workspace"
   name              = "test-workspace"
   terraform_version = "1.3.5"